Define the Head and Body of an HTML Document
Basic HTML and HTML5: Define the Head and Body of an HTML Document

You can add another level of organization in your HTML document within the html tags with the head and body elements. Any markup with information about your page would go into the head tag. Then any markup with the content of the page (what displays for a user) would go into the body tag.

Metadata elements, such as link, meta, title, and style, typically go inside the head element.

Here's an example of a page's layout:

    <!DOCTYPE html>
    <html>
      <head>
        <!-- metadata elements -->
      </head>
      <body>
        <!-- page contents -->
      </body>
    </html>

Edit the markup so there's a head and a body. The head element should only include the title, and the body element should only include the h1 and p.