Retina Image for Higher Resolution Displays
Responsive Web Design Principles: Use a Retina Image for Higher Resolution Displays

The simplest way to make your images appear "retina" (and optimize them for retina displays) is to define their width and height values as only half of what the original file is.

Here is an example of an image that is only using half of the original height and width:

    <style>
      img { height: 250px; width: 250px; }
    </style>
    <img src="coolPic500x500" alt="A most excellent picture">

Set the width and height of the img tag to half of their original values. In this case, both the original height and the original width are 200px.
Passed
Your img tag should have a width of 100 pixels.
Passed
Your img tag should have a height of 100 pixels.